reimbly/requirements.txt
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.1
pytest==7.4.3
pytest-mock==3.12.0 